Attempt at splitting the components into different files, can I write unit tests now?
